+++ packages/expo-router/CHANGELOG.md
@@ -78,6 +78,7 @@ _This version does not introduce any user-facing changes._
 
 - Bumped iOS deployment target to 15.1. ([#30840](https://github.com/expo/expo/pull/30840) by [@tsapeta](https://github.com/tsapeta))
 - Update to React Navigation v7 ([#28109](https://github.com/expo/expo/pull/28109) by [@marklawlor](https://github.com/marklawlor))
+- Remove generic from `Href` type and navigation hooks and functions
 
 ### 🎉 New features